Reorganise tests to make more logical when manually testing
